// src/api/generated/models/chatRequestMode.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ChatRequestMode = typeof ChatRequestMode[keyof typeof ChatRequestMode];


export const ChatRequestMode = {
  CODE: 'CODE',
  CHAT: 'CHAT',
  RESUME: 'RESUME',
} as const;